requirements.txt
-i https://pypi.org/simple
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
h11==0.13.0
html5lib==1.1
idna==3.3
lxml==4.9.1
numpy==1.23.1
outcome==1.2.0
pandas==1.4.3
pycparser==2.21
pyopenssl==22.0.0
pysocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
schedule==1.1.0
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.21.0
trio-websocket==0.9.2
urllib3[secure,socks]==1.26.10
webdriver-manager==3.8.2
webencodings==0.5.1
wsproto==1.1.0
